membw + restructuring
