+++ src/stubs/stubs_host2device_bandwidth.jl
