+++ src/stubs/stubs_p2p_bandwidth.jl
