+++ src/utility.jl
@@ -1,0 +1,3 @@
+function logspace(start, stop, length)
+    return exp2.(range(log2(start), log2(stop); length=length))
+end
